Added new Visual Studio Solution file
Added a new Visual Studio Solution file (`eshop-microservices.sln`) to manage projects. The solution file is set to format version 12.00 and specifies Visual Studio version 17 as the version used. The minimum required Visual Studio version to open this solution is set to 10.0.40219.1. Two global sections, `SolutionProperties` and `ExtensibilityGlobals`, are defined. The `HideSolutionNode` property in `SolutionProperties` is set to FALSE, making the solution node visible in Solution Explorer. A unique `SolutionGuid` is defined in the `ExtensibilityGlobals` section.